moving some synthesis params to an advanced configuration panel
