+++ packages/webapp/src/pages/Synthesize/DataSynthesisParameters/index.ts
@@ -1,0 +1,5 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './DataSynthesisParameters'
